+++ couchpotato/core/media/movie/searcher/main.py
@@ -183,7 +183,7 @@ class MovieSearcher(SearcherBase, MovieTypeBase):
                 found_releases += fireEvent('release.create_from_search', results, movie, quality_type, single = True)
 
                 # Try find a valid result and download it
-                if fireEvent('searcher.try_download_result', results, movie, quality_type, manual, single = True):
+                if fireEvent('release.try_download_result', results, movie, quality_type, manual, single = True):
                     ret = True
 
                 # Remove releases that aren't found anymore
